refactor: app과 index 분리
